Parsing functions are seperated
