+++ Makefile
@@ -8,7 +8,10 @@ INCLUDES = -I./includes/ -I./libft/includes/
 SRCS = main.c \
 	./tokenizing/tokenizing.c ./tokenizing/clean_tokens.c \
 	./tokenizing/print_tokens.c ./tokenizing/tokenizing_utils.c \
-	./parsing/parsing1.c ./parsing/print_ast.c
+	./parsing/parsing.c ./parsing/print_ast.c ./parsing/addArg.c \
+	./parsing/addRedir.c ./parsing/cleanAST.c ./parsing/createNode.c \
+	./parsing/isRedirection.c ./parsing/parseCommand.c \
+	./parsing/parseExpression.c ./parsing/parseRedirection.c \
 
 OBJS = $(SRCS:.c=.o)
