+++ parsing/parseCommand.c
@@ -1,0 +1,43 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   parseCommand.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jkubaev <jkubaev@student.42berlin.de>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/09/04 11:34:27 by jkubaev           #+#    #+#             */
+/*   Updated: 2025/09/04 11:41:47 by jkubaev          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/minishell.h"
+
+t_node	*parse_command(t_token **list)
+{
+	t_node	*node;
+	t_token	*tmp;
+	t_redir_token	*redir;
+	
+	tmp = *list;
+	node = createNode(COMMAND);
+	if (!node)
+		return (NULL);
+	while (tmp && tmp->type != T_PIPE && tmp->type != T_EOF)
+    {
+        if (tmp->type == T_WORD || tmp->type == T_VAR)
+        {
+            if (add_arg(node, &tmp))
+				return (freeAST(node), NULL);
+			tmp = tmp->next;
+        }
+        else if (is_redirection(tmp->type))
+        {
+            redir = parseRedirection(&tmp);
+			if (!redir)
+				return (freeAST(node), NULL);
+            add_redir(node, redir);
+        }
+    }
+	*list = tmp;
+	return (node);
+}
