+++ parsing/parsing.c
@@ -1,0 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   parsing.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jkubaev <jkubaev@student.42berlin.de>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/09/01 14:37:42 by jkubaev           #+#    #+#             */
+/*   Updated: 2025/09/04 11:35:38 by jkubaev          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/minishell.h"
+
+void	ft_parse(char **input)
+{
+	t_token	*token_list;
+	t_token	*tmp_list;
+	t_node	*nodes;
+
+	token_list = ft_tokenize(*input);
+	tmp_list = token_list;
+	nodes = parse_expression(&tmp_list);
+	if (!nodes)
+	{
+		printf("returned null\n");
+		clean_tokens(&token_list, 0);
+		return ;
+	}
+	print_ast(nodes, 0);
+	freeAST(nodes);
+	clean_tokens(&token_list, 0);
+}
