+++ readline.supp
@@ -1,0 +1,42 @@
+{
+   readline_leak_malloc_xmalloc
+   Memcheck:Leak
+   fun:malloc
+   fun:xmalloc
+   ...
+   fun:rl_initialize
+   fun:readline
+   fun:main
+}
+{
+   readline_leak_realloc
+   Memcheck:Leak
+   fun:realloc
+   fun:xrealloc
+   ...
+   fun:rl_initialize
+   fun:readline
+   fun:main
+}
+{
+   readline_leak_tgetent
+   Memcheck:Leak
+   fun:malloc
+   fun:_nc_setupterm
+   fun:tgetent_sp
+   fun:_rl_init_terminal_io
+   fun:rl_initialize
+   fun:readline
+   fun:main
+}
+{
+   readline_leak_setlocale
+   Memcheck:Leak
+   fun:malloc
+   fun:setlocale
+   fun:_rl_init_locale
+   fun:_rl_init_eightbit
+   fun:rl_initialize
+   fun:readline
+   fun:main
+}
